take_home/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2022.1
sqlparse==0.4.2
